move app screen state content to design system
