+++ design_system/src/main/java/com/sanaa/designsystem/design_system/component/screen_state_content/NetworkDisconnectionContact.kt
@@ -1,4 +1,4 @@
-package com.sanaa.presentation.screen.componants
+package com.sanaa.designsystem.design_system.component.screen_state_content
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
